// app/xmlns/www.musicxml.org.d.ts
interface BaseType {
	_exists: boolean;
	_namespace: string;
}
// ...
export interface accidental extends _accidental { constructor: { new(): accidental }; }
// ...
export interface attributes extends _attributes { constructor: { new(): attributes }; }
// ...
export interface backup extends _backup { constructor: { new(): backup }; }
// ...
export interface barline extends _barline { constructor: { new(): barline }; }
// ...
export interface beam extends _beam { constructor: { new(): beam }; }
// ...
export interface bookmark extends _bookmark { constructor: { new(): bookmark }; }
// ...
export type color = string;
// ...
export interface direction extends _direction { constructor: { new(): direction }; }
// ...
export interface dynamics extends _dynamics { constructor: { new(): dynamics }; }
// ...
export interface empty extends _empty { constructor: { new(): empty }; }
// ...
export interface emptyPlacement extends _emptyPlacement { constructor: { new(): emptyPlacement }; }
// ...
export interface figuredBass extends _figuredBass { constructor: { new(): figuredBass }; }
// ...
export interface fontSize extends _fontSize { constructor: { new(): fontSize }; }
export var fontSize: { new(): fontSize };

/** The font-style type represents a simplified version of the CSS font-style property. */
export type fontStyle = ("normal" | "italic");
interface _fontStyle extends Primitive._string { content: fontStyle; }

/** The font-weight type represents a simplified version of the CSS font-weight property. */
export type fontWeight = ("normal" | "bold");
// ...
export interface formattedText extends _formattedText { constructor: { new(): formattedText }; }
// ...
export interface forward extends _forward { constructor: { new(): forward }; }
// ...
export interface grace extends _grace { constructor: { new(): grace }; }
// ...
export interface grouping extends _grouping { constructor: { new(): grouping }; }
// ...
export interface harmony extends _harmony { constructor: { new(): harmony }; }
// ...
export interface instrument extends _instrument { constructor: { new(): instrument }; }
// ...
export interface level extends _level { constructor: { new(): level }; }
// ...
export interface link extends _link { constructor: { new(): link }; }
// ...
export interface lyric extends _lyric { constructor: { new(): lyric }; }
// ...
export interface notations extends _notations { constructor: { new(): notations }; }
// ...
interface _note extends BaseType {
	attack: number;
	color: string;
	defaultX: number;
	defaultY: number;
	dynamics: number;
	endDynamics: number;
	fontFamily: string;
	fontSize: fontSize;
	fontStyle: fontStyle;
	fontWeight: fontWeight;
	pizzicato: yesNo;
	printDot: yesNo;
	printLyric: yesNo;
	printObject: yesNo;
	printSpacing: yesNo;
	relativeX: number;
	relativeY: number;
	release: number;
	timeOnly: string;
	accidental?: accidental;
	beam?: beam[];
	/** The chord element indicates that this note is an additional chord tone with the preceding note. The duration of this note can be no longer than the preceding note. In MuseData, a missing duration indicates the same length as the previous note, but the MusicXML format requires a duration for chord notes too. */
	chord?: empty[];
	/** The cue element indicates the presence of a cue note. */
	cue: empty;
	/** One dot element is used for each dot of prolongation. The placement element is used to specify whether the dot should appear above or below the staff line. It is ignored for notes that appear on a staff space. */
	dot?: emptyPlacement[];
	/** Duration is a positive number specified in division units. This is the intended duration vs. notated duration (for instance, swing eighths vs. even eighths, or differences in dotted notes in Baroque-era music). Differences in duration specific to an interpretation or performance should use the note element's attack and release attributes. */
	duration: number[];
	footnote?: formattedText;
	grace: grace;
	instrument?: instrument;
	level?: level;
	lyric?: lyric[];
	notations?: notations[];
	notehead?: notehead;
	noteheadText?: noteheadText;
	pitch: pitch[];
	play?: play;
	rest: rest[];
	/** Staff assignment is only needed for music notated on multiple staves. Used by both notes and directions. Staff values are numbers, with 1 referring to the top-most staff in a part. */
	staff?: number;
	stem?: stem;
	tie?: tie[];
	timeModification?: timeModification;
	type?: noteType;
	unpitched: unpitched[];
	voice?: string;
}
export interface note extends _note { constructor: { new(): note }; }
// ...
export interface notehead extends _notehead { constructor: { new(): notehead }; }
// ...
export interface noteheadText extends _noteheadText { constructor: { new(): noteheadText }; }
// ...
export interface noteType extends _noteType { constructor: { new(): noteType }; }
// ...
export interface pitch extends _pitch { constructor: { new(): pitch }; }
// ...
export interface play extends _play { constructor: { new(): play }; }
// ...
export interface print extends _print { constructor: { new(): print }; }
// ...
export interface rest extends _rest { constructor: { new(): rest }; }
// ...
interface _ScorePartwiseTypePartTypeMeasureType extends BaseType {
	implicit: yesNo;
	nonControlling: yesNo;
	number: string;
	width: number;
	attributes?: attributes[];
	backup?: backup[];
	barline?: barline[];
	bookmark?: bookmark[];
	direction?: direction[];
	figuredBass?: figuredBass[];
	forward?: forward[];
	grouping?: grouping[];
	harmony?: harmony[];
	link?: link[];
	note?: note[];
	print?: print[];
	sound?: sound[];
}
export interface ScorePartwiseTypePartTypeMeasureType extends _ScorePartwiseTypePartTypeMeasureType { constructor: { new(): ScorePartwiseTypePartTypeMeasureType }; }
// ...
export interface sound extends _sound { constructor: { new(): sound }; }
// ...
export interface stem extends _stem { constructor: { new(): stem }; }
// ...
export interface tie extends _tie { constructor: { new(): tie }; }
// ...
export interface timeModification extends _timeModification { constructor: { new(): timeModification }; }
export var timeModification: { new(): timeModification };

/** The time-only type is used to indicate that a particular playback-related element only applies particular times through a repeated section. The value is a comma-separated list of positive integers arranged in ascending order, indicating which times through the repeated section that the element applies. */
export type timeOnly = string;
// ...
export interface unpitched extends _unpitched { constructor: { new(): unpitched }; }
// ...
export type yesNo = ("yes" | "no");